+++ .gitignore
@@ -4,4 +4,3 @@
 # Logs
 *.json
 !assets/*.json
-!KOTA_SURABAYA.json
