+++ cmd/entity/response.go
